Resolved CID 351658 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->socket_count.load(std::memory_order_seq_cst) to format specifier %u was expected to have type unsigned int but has type unsigned long.

+++ src/shared/Network/SocketMgrLinux.cpp
@@ -111,7 +111,7 @@ void SocketMgr::SpawnWorkerThreads()
 
 void SocketMgr::ShowStatus()
 {
-    LogDefault("sockets count = %u", socket_count.load());
+    LogDefault("sockets count = %u", static_cast<uint32_t>(socket_count.load()));
 }
 
 bool SocketWorkerThread::runThread()
